refactor(game): Organized player vs. engine game executable
