--- src/bin/test_opening_book.rs
@@ -1,10 +1,0 @@
-// NOTE: This test binary is currently disabled as UCI parsing functionality is deprecated
-// TODO: Re-implement using ChessMove directly instead of UCI strings
-
-fn main() {
-    println!("=== Opening Book Test ===");
-    println!();
-    println!("This test is currently disabled.");
-    println!("UCI parsing functionality has been deprecated.");
-    println!("Please re-implement opening book tests using ChessMove directly.");
-}
