--- src/bin/two_player_chess.rs
@@ -1,6 +1,0 @@
-use rusty_chess::game::Game;
-
-fn main() {
-    let mut game = Game::new();
-    game.run();
-}
