+++ src/engine/mod.rs
@@ -1,0 +1,3 @@
+pub mod game;
+
+pub use game::{AiGame, PlayerAction, parse_square, square_to_notation};
