+++ src/terminal/introduction.rs
@@ -1,0 +1,31 @@
+pub fn print_introduction() {
+    println!(
+        r#"
+    ╔═══════════════════════════════════════════════════════════════╗
+    ║                                                               ║
+    ║   ♜ ♞ ♝ ♛ ♚ ♝ ♞ ♜      ____  _   _ ____ _______   __          ║
+    ║   ♟ ♟ ♟ ♟ ♟ ♟ ♟ ♟     |  _ \| | | / ___|_   _\ \ / /          ║
+    ║   . . . . . . . .     | |_) | | | \___ \ | |  \ V /           ║
+    ║   . . . . . . . .     |  _ <| |_| |___) || |   | |            ║
+    ║   . . . . . . . .     |_| \_\\___/|____/ |_|   |_|            ║
+    ║   . . . . . . . .                                             ║
+    ║   ♙ ♙ ♙ ♙ ♙ ♙ ♙ ♙       ____ _   _ _____ ____ ____            ║
+    ║   ♖ ♘ ♗ ♕ ♔ ♗ ♘ ♖      / ___| | | | ____/ ___/ ___|           ║
+    ║                       | |   | |_| |  _| \___ \___ \           ║
+    ║                       | |___|  _  | |___ ___) |__) |          ║
+    ║                        \____|_| |_|_____|____/____/           ║
+    ║                                                               ║
+    ╠═══════════════════════════════════════════════════════════════╣
+    ║                                                               ║
+    ║          A terminal chess engine written in Rust              ║
+    ║                                                               ║
+    ║    Features:                                                  ║
+    ║      - Minimax search with alpha-beta pruning                 ║
+    ║      - Quiescence search for tactical accuracy                ║
+    ║      - Opening book support                                   ║
+    ║      - Full move validation and game state tracking           ║
+    ║                                                               ║
+    ╚═══════════════════════════════════════════════════════════════╝
+"#
+    );
+}
