+++ src/terminal/mod.rs
@@ -1,0 +1,7 @@
+mod instructions;
+mod introduction;
+mod settings;
+
+pub use instructions::print_instructions;
+pub use introduction::print_introduction;
+pub use settings::{ChessEngineSettings, DisplaySettings, get_chess_engine_settings};
